refactoring: 절대경로 변경 반영하기
