+++ src/App.jsx
@@ -2,7 +2,7 @@ import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 
 import Header from '@/components/organism/Header/Header';
 import Main from './pages/Main/Main';
-import TrainTransfer from './pages/TrainTransfer/main';
+import TrainTransfer from './pages/TrainTransfer/TrainTransfer';
 // import TransferResult from 'pages/TrainTransferResult/main';
 // import BusSchedulePage from 'pages/BusSchedule/main';
