+++ src/pages/TrainTransfer/components/PageExplain/PageExplain.css
@@ -1,0 +1,18 @@
+.info-container {
+  height: 100px;
+  min-height: 150px;
+  display: flex;
+  line-height: 1.2;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  font-size: 20px;
+}
+
+.info-container .point {
+  font-weight: bold;
+}
+
+.info-container .point-content {
+  color: #0047ff;
+}
