+++ src/pages/TrainTransfer/components/PageExplain/PageExplain.jsx
@@ -1,0 +1,16 @@
+const PageExplain = () => {
+  return (
+    <div className="info-container">
+      <div className="info-content">
+        <div>
+          <span className="point-content">열차</span>와{' '}
+          <span className="point-content">버스</span>
+        </div>
+        <p>시간표를</p>
+        <p>한 눈에 확인해 보세요</p>
+      </div>
+    </div>
+  );
+};
+
+export default PageExplain;
