+++ src/pages/TrainTransfer/hooks/useFormStations.js
@@ -1,119 +1,119 @@
 import { useEffect, useState } from 'react';
-import useDebounce from 'hooks/useDebounce';
-import STATIONS from 'assets/json/Station';
+import useDebounce from '@/hooks/useDebounce';
+import STATIONS from '@/assets/json/Station';
 
 const useFormStations = ({
-	departStation,
-	setDepartStation,
-	arriveStation,
-	setArriveStation,
+  departStation,
+  setDepartStation,
+  arriveStation,
+  setArriveStation,
 }) => {
-	const { run } = useDebounce();
-	const [departToggle, setDepartToggle] = useState(false);
-	const [arriveToggle, setArriveToggle] = useState(false);
-	const [stationLists, setStationLists] = useState([]);
-	const [searchInputValue, setSearchInputValue] = useState('');
+  const { run } = useDebounce();
+  const [departToggle, setDepartToggle] = useState(false);
+  const [arriveToggle, setArriveToggle] = useState(false);
+  const [stationLists, setStationLists] = useState([]);
+  const [searchInputValue, setSearchInputValue] = useState('');
 
-	const handleChangeInputValue = (event) => {
-		const { value } = event.target;
+  const handleChangeInputValue = (event) => {
+    const { value } = event.target;
 
-		setSearchInputValue(value);
-	};
+    setSearchInputValue(value);
+  };
 
-	const handleReverseStation = () => {
-		setDepartStation(arriveStation);
-		setArriveStation(departStation);
-	};
+  const handleReverseStation = () => {
+    setDepartStation(arriveStation);
+    setArriveStation(departStation);
+  };
 
-	const handleClickDepartStation = () => {
-		setDepartToggle(!departToggle);
-		setArriveToggle(false);
-	};
+  const handleClickDepartStation = () => {
+    setDepartToggle(!departToggle);
+    setArriveToggle(false);
+  };
 
-	const handleClickArriveStation = () => {
-		setArriveToggle(!arriveToggle);
-		setDepartToggle(false);
-	};
+  const handleClickArriveStation = () => {
+    setArriveToggle(!arriveToggle);
+    setDepartToggle(false);
+  };
 
-	const handleChangeStation = (event) => {
-		const li = event.target.closest('li');
+  const handleChangeStation = (event) => {
+    const li = event.target.closest('li');
 
-		if (!li) return;
+    if (!li) return;
 
-		const { innerText } = li;
-		if (departToggle) {
-			setDepartStation(innerText);
-		}
+    const { innerText } = li;
+    if (departToggle) {
+      setDepartStation(innerText);
+    }
 
-		if (arriveToggle) {
-			setArriveStation(innerText);
-		}
-	};
+    if (arriveToggle) {
+      setArriveStation(innerText);
+    }
+  };
 
-	const isIncludeKeyword = (station, keyword) => {
-		keyword = '^' + keyword;
-		const reg = new RegExp(keyword);
+  const isIncludeKeyword = (station, keyword) => {
+    keyword = '^' + keyword;
+    const reg = new RegExp(keyword);
 
-		return reg.test(station) ? true : false;
-	};
+    return reg.test(station) ? true : false;
+  };
 
-	const makeStationLists = (keyword) => {
-		const stationLists = [];
+  const makeStationLists = (keyword) => {
+    const stationLists = [];
 
-		if (keyword === '') {
-			setStationLists(stationLists);
-			return;
-		}
+    if (keyword === '') {
+      setStationLists(stationLists);
+      return;
+    }
 
-		STATIONS.forEach((station) => {
-			if (isIncludeKeyword(station, keyword)) {
-				stationLists.push(station);
-			}
-		});
+    STATIONS.forEach((station) => {
+      if (isIncludeKeyword(station, keyword)) {
+        stationLists.push(station);
+      }
+    });
 
-		setStationLists(stationLists);
-	};
+    setStationLists(stationLists);
+  };
 
-	const handleSearchStationByInputValue = (event) => {
-		const { value } = event.target;
+  const handleSearchStationByInputValue = (event) => {
+    const { value } = event.target;
 
-		run(() => {
-			makeStationLists(value);
-		});
-	};
+    run(() => {
+      makeStationLists(value);
+    });
+  };
 
-	useEffect(() => {
-		if (departStation !== '밀양') {
-			setArriveStation('밀양');
-			setDepartToggle(false);
-			setStationLists([]);
-		}
-	}, [departStation, setArriveStation]);
+  useEffect(() => {
+    if (departStation !== '밀양') {
+      setArriveStation('밀양');
+      setDepartToggle(false);
+      setStationLists([]);
+    }
+  }, [departStation, setArriveStation]);
 
-	useEffect(() => {
-		if (arriveStation !== '밀양') {
-			setDepartStation('밀양');
-			setArriveToggle(false);
-			setStationLists([]);
-		}
-	}, [arriveStation, setDepartStation]);
+  useEffect(() => {
+    if (arriveStation !== '밀양') {
+      setDepartStation('밀양');
+      setArriveToggle(false);
+      setStationLists([]);
+    }
+  }, [arriveStation, setDepartStation]);
 
-	return {
-		states: {
-			departToggle,
-			arriveToggle,
-			stationLists,
-			searchInputValue,
-		},
-		methods: {
-			handleChangeInputValue,
-			handleReverseStation,
-			handleClickDepartStation,
-			handleClickArriveStation,
-			handleSearchStationByInputValue,
-			handleChangeStation,
-		},
-	};
+  return {
+    states: {
+      departToggle,
+      arriveToggle,
+      stationLists,
+      searchInputValue,
+    },
+    methods: {
+      handleChangeInputValue,
+      handleReverseStation,
+      handleClickDepartStation,
+      handleClickArriveStation,
+      handleSearchStationByInputValue,
+      handleChangeStation,
+    },
+  };
 };
 
 export default useFormStations;
